[WIP] Progress towards full enumeration of HTML elements
This uses the combination of h/ref and h/alt to store the definitions of
many HTML elements in a single model. The models for particular elements
can be accessed by entries in the :bindings map of the model.

At least a few of the nested recursive specs work on toy data. An encouraging sign!