Sketch of new parsing method with error handling/propagation
The general idea is to let the rendering and evaluation parts of the
pipeline surface errors in context - potentially even inserting them
into the rendered page. This method also allows for multiple errors in a
page to be identified sequentially, as opposed to merely stopping on the
first one.

This was inspired by the approach taken by clojure.java.shell/sh.